Rename: domain 폴더 생성 및 CarTest/CarsTest 이동
