+++ src/test/java/racingcar/model/domain/CarTest.java
@@ -1,4 +1,4 @@
-package racingcar.model;
+package racingcar.model.domain;
 
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
