+++ src/test/java/racingcar/model/domain/CarsTest.java
@@ -1,4 +1,4 @@
-package racingcar.model;
+package racingcar.model.domain;
 
 import camp.nextstep.edu.missionutils.test.NsTest;
 import org.junit.jupiter.api.BeforeEach;
